Moving the LoggedInUser and the DAO
