+++ core/src/androidTest/java/io/plaidapp/core/designernews/data/login/LoginLocalDataSourceTest.kt
@@ -18,7 +18,7 @@ package io.plaidapp.core.designernews.data.login
 
 import android.content.Context
 import android.support.test.InstrumentationRegistry.getInstrumentation
-import io.plaidapp.core.designernews.data.users.model.LoggedInUser
+import io.plaidapp.core.designernews.data.login.model.LoggedInUser
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
