+++ core/src/main/java/io/plaidapp/core/data/database/Converters.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.plaidapp.core.data
+package io.plaidapp.core.data.database
 
 import android.arch.persistence.room.TypeConverter
