+++ core/src/main/java/io/plaidapp/core/designernews/data/poststory/PostStoryService.java
@@ -32,8 +32,7 @@ import io.plaidapp.core.designernews.data.api.DesignerNewsService;
 import io.plaidapp.core.designernews.data.login.LoginRepository;
 import io.plaidapp.core.designernews.data.poststory.model.NewStoryRequest;
 import io.plaidapp.core.designernews.data.stories.model.Story;
-import io.plaidapp.core.designernews.data.users.model.LoggedInUser;
-import io.plaidapp.core.designernews.data.users.model.User;
+import io.plaidapp.core.designernews.data.login.model.LoggedInUser;
 import io.plaidapp.core.designernews.data.stories.model.StoryKt;
 import retrofit2.Call;
 import retrofit2.Response;
