+++ designernews/src/main/java/io/plaidapp/designernews/ui/story/StoryActivity.java
@@ -65,8 +65,7 @@ import io.plaidapp.core.data.Result;
 import io.plaidapp.core.designernews.Injection;
 import io.plaidapp.core.designernews.data.login.LoginRepository;
 import io.plaidapp.core.designernews.data.stories.model.Story;
-import io.plaidapp.core.designernews.data.users.model.LoggedInUser;
-import io.plaidapp.core.designernews.data.users.model.User;
+import io.plaidapp.core.designernews.data.login.model.LoggedInUser;
 import io.plaidapp.core.designernews.domain.model.Comment;
 import io.plaidapp.core.ui.transitions.GravityArcMotion;
 import io.plaidapp.core.ui.transitions.MorphTransform;
