+++ designernews/src/test/java/io/plaidapp/designernews/ui/login/LoginViewModelTest.kt
@@ -24,7 +24,7 @@ import com.nhaarman.mockito_kotlin.whenever
 import io.plaidapp.R
 import io.plaidapp.core.data.Result
 import io.plaidapp.core.designernews.data.login.LoginRepository
-import io.plaidapp.core.designernews.data.users.model.LoggedInUser
+import io.plaidapp.core.designernews.data.login.model.LoggedInUser
 import io.plaidapp.core.util.event.Event
 import io.plaidapp.test.shared.LiveDataTestUtil
 import io.plaidapp.test.shared.provideFakeCoroutinesContextProvider
